+++ xbmc/pictures/GUIWindowSlideShow.cpp
@@ -641,7 +641,7 @@ void CGUIWindowSlideShow::Render()
 
   if (m_slides.at(m_iCurrentSlide)->IsVideo())
   {
-    g_graphicsContext.SetViewWindow(0, 0, m_coordsRes.iWidth, m_coordsRes.iHeight);
+    g_graphicsContext.SetViewWindow(0, 0, static_cast<float>(m_coordsRes.iWidth), static_cast<float>(m_coordsRes.iHeight));
     g_graphicsContext.SetRenderingResolution(g_graphicsContext.GetVideoResolution(), false);
     g_application.m_pPlayer->Render(true, 255);
     g_graphicsContext.SetRenderingResolution(m_coordsRes, m_needsScaling);
